consolidate generate user prompt to the legacy file
